Add context check; move task nodes to separate file
